# api/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Error:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  dto.Address:
    properties:
      city:
        type: string
      country:
        type: string
      street:
        type: string
    type: object
  dto.Client:
    properties:
      birthday:
        type: string
      city:
        type: string
      country:
        type: string
      gender:
        type: string
      name:
        type: string
      street:
        type: string
      surname:
        type: string
    required:
    - birthday
    - gender
    - name
    - surname
    type: object
...
host: aboba.com
info:
  contact: {}
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger CRUD Home appliance store API
  version: "1.0"
paths:
  /api/v1/clients:
    get:
      consumes:
      - application/json
      description: That endpoint retrieve all registered client in system
      parameters:
      - description: limit get data
        in: query
        name: limit
        type: integer
      - description: offset get data
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Client'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get all client
      tags:
      - clients
    post:
      consumes:
      - application/json
      description: 'Client created from JSON or XML, for create endpoint required:
        name, surname, birthday, gender, address_id'
      parameters:
      - description: Client name
        in: path
        name: name
        required: true
        type: string
      - description: Client surname
        in: path
        name: surname
        required: true
        type: string
      - description: Client birthday
        in: path
        name: birthday
        required: true
        type: string
      - description: Client gender
        in: path
        name: gender
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Create client
      tags:
      - clients
  /api/v1/clients/{id}:
    delete:
      consumes:
      - application/json
      description: That methods deleting registered client in system by id
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Delete client from system
      tags:
      - clients
    patch:
      consumes:
      - application/json
      description: That endpoint update client data (change address on client)
      parameters:
      - description: Change address
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/dto.Address'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Update address on client
      tags:
      - clients
  /api/v1/clients/search:
    get:
      consumes:
      - application/json
      description: That endpoint retrieve all required registered client in system
        with gived name and surname
      parameters:
      - description: client name
        in: query
        name: name
        required: true
        type: string
      - description: client surname
        in: query
        name: surname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Client'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get client filtered by name and surname
      tags:
      - clients
...
swagger: "2.0"